CompraActivity funcionando completamente, empezando con el DeshacerActivity

+++ app/src/main/java/com/tute/estacionamientofrlp/Constantes.java
@@ -9,6 +9,6 @@ public class Constantes {
     public static String cSaldo, cUid, cCompSem, cPosSpi, sem1, sem2, pridiasem,ultdiasem,semana, apellido, nombre;
     public static ArrayList cCod;
     public static Integer montoSum;
-    public static int precioticket = 3;
+    public static double precioticket = 3;
 
 }
